moved JMeterResponseCollector to services package
Added doc comments
