+++ src/main/java/com/revature/models/Docutest.java
@@ -2,6 +2,9 @@ package com.revature.models;
 
 import java.util.List;
 
+/**
+ * Model representing a single specification file upload. Data for load testing is embedded in each Request object.
+ */
 public interface Docutest {
     
     public int getId();
