+++ src/main/java/com/revature/services/ResultSummaryService.java
@@ -8,6 +8,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+/**
+ * Service for interfacing with the ResultSummaryRepository
+ */
 @Service
 @Transactional
 public class ResultSummaryService {
