+++ src/main/java/com/revature/services/SwaggerSummaryService.java
@@ -19,6 +19,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+/**
+ * Service for interfacing with the SwaggerSummaryRepository
+ */
 @Service
 @Transactional
 public class SwaggerSummaryService {
